CHORE: refactored MyBaskets component
- Renamed component to MyBaskets because Basket vs Baskets component was
confusing.
- Internally implements MenuList instead of List.
- Some styling updates.
