--- frontend/src/components/Baskets/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./Baskets";
