+++ frontend/src/components/MyBaskets/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./MyBaskets";
